+++ lib/src/_base.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'dart:ffi';
+import 'dart:typed_data';
 
 import 'package:ffi/ffi.dart';
 import 'package:win32/win32.dart';
@@ -10,16 +11,40 @@ import 'package:win32/win32.dart';
 import 'attribute.dart';
 
 // The base object for metadata objects.
-class TokenObject {
+abstract class TokenObject {
   final IMetaDataImport2 reader;
   final int token;
 
   const TokenObject(this.reader, this.token);
 }
 
-class AttributeObject extends TokenObject {
+/// Represents an object that has attributes associated with it.
+abstract class AttributeObject extends TokenObject {
   const AttributeObject(IMetaDataImport2 reader, int token)
       : super(reader, token);
+
+  Uint8List attributeByName(String attrName) {
+    final szName = attrName.toNativeUtf16();
+    final ppData = calloc<IntPtr>();
+    final pcbData = calloc<Uint32>();
+    try {
+      final hr =
+          reader.GetCustomAttributeByName(token, szName, ppData, pcbData);
+      if (SUCCEEDED(hr)) {
+        print(
+            'GetCustomAttributeByName(${szName.toDartString()}) returned ${pcbData.value} bytes of data');
+        final sigList =
+            Pointer<Uint8>.fromAddress(ppData.value).asTypedList(pcbData.value);
+        return sigList;
+      } else {
+        throw WindowsException(hr);
+      }
+    } finally {
+      calloc.free(szName);
+      calloc.free(ppData);
+      calloc.free(pcbData);
+    }
+  }
 
   /// Enumerate all attributes that this parameter has.
   List<Attribute> get attributes {
@@ -40,8 +65,8 @@ class AttributeObject extends TokenObject {
       }
       return attributes;
     } finally {
-      reader.CloseEnum(phEnum.address);
-
+      reader.CloseEnum(phEnum.value);
+      calloc.free(phEnum);
       calloc.free(rAttrs);
       calloc.free(pcAttrs);
     }